requirements.txt
absl-py==0.15.0
altair==4.1.0
argcomplete==2.0.0
argh==0.26.2
arrow==1.2.2
asciitree==0.3.3
astropy==4.1
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
binaryornot==0.4.4
bioframe==0.2.0
biomart==0.9.2
biopython==1.79
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorlog==6.6.0
cookiecutter==1.7.3
cooler==0.8.11
cooltools==0.4.1
cycler==0.11.0
Cython==0.29.28
cytoolz==0.10.1
dataclasses==0.8
decorator==4.4.2
deprecation==2.1.0
descartes==1.1.0
dill==0.3.4
dm-sonnet==2.0.0
dm-tree==0.1.7
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gffutils==0.10.1
google-api-core==2.7.2
google-auth==2.6.5
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras-Preprocessing==1.1.2
kipoi==0.8.2
kipoi-conda==0.3.1
kipoi-utils==0.7.5
kipoiseq==0.5.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-venn==0.11.7
multiprocess==0.70.12.2
natsort==8.1.0
nest-asyncio==1.5.5
networkx==2.5.1
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
parso==0.7.1
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.7.0
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
py==1.11.0
py4j==0.10.9.3
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybedtools==0.9.0
pyBigWig==0.3.18
pycellbase==4.9.0
pyfaidx==0.6.4
Pygments==2.11.2
pypairix==0.3.7
pyparsing==3.0.8
pyrsistent==0.18.0
pysam==0.19.0
pyspark==3.2.1
pytabix==0.1
pytest==7.0.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
related==0.7.2
requests==2.27.1
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
selene-sdk==0.5.1
shapley==1.0.3
simplejson==3.17.6
six==1.15.0
sortedcontainers==2.4.0
statsmodels==0.12.2
tables==3.7.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0
tinydb==4.7.0
tomli==1.2.3
toolz==0.11.2
torch==1.9.0
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing_extensions==4.1.1
UpSetPlot==0.6.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
